+++ src/SprykerCommunity/Glue/WebhookProcessor/Plugin/GlueBackendApiApplication/WebhookProcessorResourcePlugin.php
@@ -1,0 +1,64 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * This file is part of the Spryker Commerce OS.
+ * For full license information, please view the LICENSE file that was distributed with this source code.
+ */
+
+namespace SprykerCommunity\Glue\WebhookProcessor\Plugin\GlueBackendApiApplication;
+
+use Generated\Shared\Transfer\GlueResourceMethodCollectionTransfer;
+use Generated\Shared\Transfer\GlueResourceMethodConfigurationTransfer;
+use Spryker\Glue\GlueApplication\Plugin\GlueApplication\Backend\AbstractResourcePlugin;
+use SprykerCommunity\Glue\WebhookProcessor\Controller\WebhookProcessorResourceController;
+
+/**
+ * @method \SprykerCommunity\Glue\WebhookProcessor\WebhookProcessorFactory getFactory()
+ */
+class WebhookProcessorResourcePlugin extends AbstractResourcePlugin
+{
+    protected const string RESOURCE_TYPE = 'webhook-processor';
+
+    /**
+     * {@inheritDoc}
+     *
+     * @api
+     *
+     * @return string
+     */
+    public function getType(): string
+    {
+        return static::RESOURCE_TYPE;
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @api
+     *
+     * @return string
+     */
+    public function getController(): string
+    {
+        return WebhookProcessorResourceController::class;
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @api
+     *
+     * @return \Generated\Shared\Transfer\GlueResourceMethodCollectionTransfer
+     */
+    public function getDeclaredMethods(): GlueResourceMethodCollectionTransfer
+    {
+        return (new GlueResourceMethodCollectionTransfer())
+            ->setPost(
+                (new GlueResourceMethodConfigurationTransfer())
+                    ->setAction('postAction')
+                    ->setIsProtected(false),
+            );
+    }
+}